requirements.txt
# Core dependencies for Induct Downtime Monitoring System
# Compatible with Python 3.7.16 and older OpenSSL versions

# HTTP requests with urllib3 compatibility for older systems
requests>=2.25.0,<2.28.0
urllib3>=1.26.0,<2.0.0

# HTML parsing
beautifulsoup4>=4.9.0

# Task scheduling
schedule>=1.1.0

# YAML configuration support
PyYAML>=5.4.0,<7.0.0

# Data handling (included with Python 3.7.16)
# sqlite3 - included with Python standard library
# csv - included with Python standard library
# pathlib - included with Python 3.4+
# typing - included with Python 3.5+

# Note: Removed unused dependencies (selenium, pandas, sqlalchemy, python-dateutil)
# to minimize compatibility issues and installation time